feat: Add user management features to admin portal
This commit introduces several new features to the admin portal:

- Admins can now change any user's password directly from the user list in `admin.html`. A new API endpoint `PUT /api/users/:id/password` was added to support this.

- Admins can export the full list of users to a CSV file. This is handled by a new API endpoint `GET /api/users/export` and the `json2csv` library. The `imageUrl` column has been excluded from the export as requested.

- Users can now have profile images. The User model was updated to include an `imageUrl` field. The user creation form in `admin.html` now includes a file upload for the image, and the image is displayed in the user list and in the main application header after login.
