+++ index.html
@@ -674,9 +674,6 @@ h4[onclick] {
 
         <!-- Landing Page: Four Survey Cards -->
         <div id="landingPage" class="section hidden">
-            <div id="offline-data-warning" class="hidden" style="background-color: var(--lagos-yellow); color: var(--lagos-navy); padding: 15px; border-radius: 10px; margin-bottom: 20px; text-align: center; border: 2px solid var(--lagos-red);">
-                <strong>Data not available offline.</strong> Please connect to the internet to download the required school and LGEA data.
-            </div>
             <div class="landing-header">
                 <img src="subeb.jpg" alt="LASUBEB Logo" class="landing-logo">
                 <h1 style="color:white;font-size:2.2rem; margin-left: 20px; flex-grow: 1; text-align: center;">Welcome to LASUBEB/PIMU Needs Assessment Portal</h1>
@@ -694,7 +691,6 @@ h4[onclick] {
                 <div id="userInfo" style="color: white; margin-right: 15px; display: flex; align-items: center; gap: 8px; font-size: 1rem;">
                     <img id="user-icon" src="" alt="User Icon" style="width: 40px; height: 40px; border-radius: 50%; object-fit: cover;">
                     <span id="loggedInUsername"></span>
-
                 </div>
 				<button class="btn btn-secondary" style="width:auto;padding:8px 16px;" onclick="logout()">Sign Out</button>
             </div>
@@ -733,7 +729,7 @@ h4[onclick] {
                     <h2>View Survey Reports</h2>
                     <p>Dashboard for all submitted data</p>
                 </div>
-                <div class="survey-card disabled-card" id="admin-link">
+                <div class="survey-card hidden" id="admin-link" onclick="window.location.href='admin.html'">
                     <h2>User Administration</h2>
                     <p>Manage user accounts</p>
                 </div>
@@ -4011,7 +4007,6 @@ select.form-control option {
         let forceOffline = false;
         
 // Combined window.onload function
-// Combined window.onload function
 window.onload = async function() {
     try {
         forceOffline = JSON.parse(localStorage.getItem('auditAppForceOffline')) || false;
@@ -4026,7 +4021,6 @@ window.onload = async function() {
 
     console.log('App initialized');
     await loadAllData(); // Load all CSV data from dropdown.js
-    checkOfflineDataStatus();
     loadSession();
 
     if (currentUser) {
@@ -4037,6 +4031,7 @@ window.onload = async function() {
         updateDashboard();
         loadRecords();
         updateConnectionStatus();
+        updateAdminLinkVisibility(); // Check role on session load
     } else {
         document.getElementById('authScreen').classList.remove('hidden');
         document.getElementById('mainApp').classList.add('hidden');
@@ -4075,13 +4070,11 @@ function updateUserInfo() {
     }
 }
 
-function updateAdminLinkState() {
-    const adminLink = document.getElementById('admin-link');
-    if (!adminLink) return;
 
-    let user = null;
+function loadSession() {
     try {
         const savedUser = localStorage.getItem('auditAppCurrentUser');
+
         if (savedUser) {
             const user = JSON.parse(savedUser);
             currentUser = user.username;
@@ -4098,41 +4091,39 @@ function updateAdminLinkState() {
             updateUserInfo();
 
             console.log('Session restored for user:', currentUser);
-
-        }
-    } catch (e) {
-        console.error('Error parsing user from localStorage', e);
-    }
-
-    if (user && user.role && user.role.trim().toLowerCase() === 'admin') {
-        // Is admin: show and enable
-        adminLink.classList.remove('hidden');
-        adminLink.classList.remove('disabled-card');
-        adminLink.onclick = () => { window.location.href = 'admin.html'; };
-    } else {
-        // Is not admin, or not logged in: hide and disable
-        adminLink.classList.add('hidden');
-        adminLink.classList.add('disabled-card');
-        adminLink.onclick = null;
-
-    }
-}
-
-function loadSession() {
-    try {
-        const savedUser = localStorage.getItem('auditAppCurrentUser');
-        if (savedUser) {
-            const user = JSON.parse(savedUser);
-            currentUser = user.username;
-            updateUserInfo();
-            console.log('Session restored for user:', currentUser);
         }
     } catch (e) {
         console.error('Error loading session:', e);
-        currentUser = null;
     }
-    // Always update the admin link state after trying to load a session.
-    updateAdminLinkState();
+}
+
+
+function updateAdminLinkVisibility() {
+    try {
+        const savedUser = localStorage.getItem('auditAppCurrentUser');
+        const adminLink = document.getElementById('admin-link');
+
+        if (!adminLink) return; // Exit if the link isn't on the page
+
+        if (savedUser) {
+            const user = JSON.parse(savedUser);
+            // Show admin link only if user role is 'admin'
+            if (user.role === 'admin') {
+                adminLink.classList.remove('hidden');
+            } else {
+                adminLink.classList.add('hidden');
+            }
+        } else {
+            // If no user is logged in, always hide the admin link
+            adminLink.classList.add('hidden');
+        }
+    } catch (e) {
+        console.error('Error updating admin link visibility:', e);
+        const adminLink = document.getElementById('admin-link');
+        if (adminLink) {
+            adminLink.classList.add('hidden');
+        }
+    }
 }
 
 document.addEventListener('DOMContentLoaded', function() {
@@ -4162,13 +4153,7 @@ async function login() {
 
             // Update user-specific UI elements on the landing page
             updateUserInfo();
-
-            loadLocalGovernments();
-            updateDashboard();
-            loadRecords();
-            updateConnectionStatus();
-            updateAdminLinkState(); // Update link state after login
-
+            updateAdminLinkVisibility();
         } else {
             let msg = document.getElementById('loginErrorMsg');
             if (!msg) {
@@ -4344,21 +4329,6 @@ function removeFile(button) {
 
 // This function is no longer needed as we are not using a global uploadedFiles array
 // function updateFileDisplay() {}
-
-function checkOfflineDataStatus() {
-    const warningDiv = document.getElementById('offline-data-warning');
-    if (!warningDiv) return;
-
-    // Check if the main data source is empty and if the app is offline.
-    const isDataMissing = Object.keys(lagosStateData).length === 0;
-    const offline = !isEffectivelyOnline();
-
-    if (isDataMissing && offline) {
-        warningDiv.classList.remove('hidden');
-    } else {
-        warningDiv.classList.add('hidden');
-    }
-}
 
 // Add this anywhere in your <script> tag
 function toggleMobileSidebar() {
